requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.1
MarkupSafe==3.0.2
PyMuPDF==1.26.4
python-docx==1.2.0
typing_extensions==4.15.0
Werkzeug==3.1.3
